requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.2
Pillow==10.1.0
pydantic[dotenv]==2.5.0
pydantic-settings
tenacity==8.2.3

# Для разработки
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
